requirements.txt
pandas
flask
scikit-learn
pyspellchecker
tqdm
numpy